Show text when random spinning ends
Now I wanted to do this nicely by using the built in promises supplied
by the library, but there is no nice way of waiting on the all the
promises in the child components. Yes I know about Promise.all(), but
the coding that was just too smelly.

This way isn't technically linked to when it actualy finishes, but
rather just uses the max duration supplied; I think it's close enough,
and I'll add a transition anyways to hide that even more.